requirement.txt
streamlit
langchain
langchain-groq
python-dotenv
streamlit-oauth
supabase